------------------------------------------------------------------------ r10 | poftwaresatent | 2008-03-25 23:04:35 +0100 (Tue, 25 Mar 2008) | 22 lines Changed paths:    A /trunk/haiko    A /trunk/haiko/Makefile    A /trunk/haiko/gltrackball.c    A /trunk/haiko/gltrackball.h    A /trunk/haiko/haiko.c    A /trunk/haiko/haiko.conf    A /trunk/haiko/tball.c    A /trunk/haiko/tokenize.l    A /trunk/haiko/tparse.c    A /trunk/haiko/trackball.c    A /trunk/haiko/trackball.h    A /trunk/haiko/voxel.c    A /trunk/haiko/voxel.h    A /trunk/haiko/voxel_parse.y    A /trunk/haiko/wrap_gl.h    A /trunk/haiko/wrap_glut.h
Haiko is THE minimalist pixel dog. The Haiko application plots it on
the screen... but the data is stored in a voxel-oriented config file,
so you can show any sort of volume data.

This code is released under a BSD style license. The trackball and
gltrackball code, which have similar licenses, was copied from
xscreensaver-5.05.

At the moment, the Makefile will only work on (my) Apple OS X, but the
code should run on any system with OpenGL and GLUT support. A mid-term
goal is to add Haiko to the xscreensaver collection, which is the main
reason that this is pure C (not sure if it's ANSI compliant though).

Given that the config file parser is written using Lex and Yacc (and
this is the first time I've played with those tools), it might present
compatibiity issues there as well, for example with Flex / Bison on
GNU systems.

Oh, yeah, and the TO-DO list is pretty long still, like figuring out
proper lighting, permuting coordinate axes, adding visual effects like
pulsating voxels or whatever. A little pet project...

------------------------------------------------------------------------